Implement some struct parsing

+++ tests/feature_tests/error_declaration.c
@@ -5,16 +5,16 @@ int main() {
   // error: missing identifier name in declaration
   int *;
 
-  // error: two or more data types in declaration specifiers
-  int long a;
+  // error: unrecognized set of type specifiers
+  int int a;
 
-  // error: both signed and unsigned in declaration specifiers
+  // error: unrecognized set of type specifiers
   unsigned signed int a;
 
   // error: extern variable has initializer
   extern int a = 10;
 
-  // error: two or more storage classes in declaration specifiers
+  // error: too many storage classes in declaration specifiers
   extern auto int b;
 
   {
